Modifying configuration interface
In order to free indexers from having to deal with json and other
"formatting" issue, a little refactoring must be made. Essentially the
configuration format will move out of the scope of the indexer (indexers
are jerks and probably don't care about how the configuration is stored)
and probably land in IndexerConfigurationService, although at this point
I'm not sure. But the indexer's interface will more likely deal with
valid ConfigurationData objects rather than passing jsons back and
forth...

Hope I'm not breaking everything, although I have a feeling that I'm
increasing the penises in the codebase...